+++ spec/ruby/core/process/status/wait_spec.rb
@@ -26,10 +26,18 @@ ruby_version_is "3.0" do
         status.pid.should == pid
       end
 
-      it "should not set $? to a Process::Status" do
+      it "should not set $? to the Process::Status" do
         pid = Process.spawn(ruby_cmd('exit'))
+        status = Process::Status.wait
+        $?.should_not equal(status)
+      end
+
+      it "should not change the value of $?" do
+        pid = Process.spawn(ruby_cmd('exit'))
+        Process.wait
+        status = $?
         Process::Status.wait
-        $?.should be_nil
+        status.should equal($?)
       end
 
       it "waits for any child process if no pid is given" do